feat: #33 메타데이터 추가
