+++ src/app/(client)/challenges/post/page.tsx
@@ -1,29 +1,18 @@
-'use client';
+import ChallengePostPage from '@/components/features/challenges/post/challenge-post-page';
+import { Metadata } from 'next';
 
-import ChallengePostSelector from '@/components/features/challenges/post/challenge-post-selector';
-import ChallengePostImageUploader from '@/components/features/challenges/post/challenge-post-image-uploader';
-import ChallengePostInput from '@/components/features/challenges/post/challenge-post-input';
-import ChallengePostButtonGroup from '@/components/features/challenges/post/challenge-post-button-group';
-import { useChallengeForm } from '@/lib/hooks/use-challenge-form';
-
-const ChallengePostPage: React.FC = () => {
-  const { challenge, challengeImageFile, setters, handleChange } = useChallengeForm();
-
-  return (
-    <section className="mx-auto mb-6 max-w-[320px] p-6 md:max-w-[640px]">
-      <h1 className="mb-6 text-2xl font-bold">챌린지 생성</h1>
-      <ChallengePostInput challenge={challenge} handleChange={handleChange} />
-
-      <section className="mb-6 grid grid-cols-1 gap-6 md:grid-cols-2">
-        <ChallengePostSelector challenge={challenge} setters={setters} />
-        <ChallengePostImageUploader setters={setters} challenge={challenge} />
-      </section>
-
-      <div className="flex justify-center gap-6">
-        <ChallengePostButtonGroup challenge={challenge} challengeImageFile={challengeImageFile} isEditMode={false} />
-      </div>
-    </section>
-  );
+export const metadata: Metadata = {
+  title: '챌린지 생성 | DOGATHER',
+  description: '새로운 챌린지를 생성하는 페이지입니다.',
+  openGraph: {
+    title: '챌린지 생성 | DOGATHER',
+    description: '새로운 챌린지를 생성하는 페이지입니다.',
+    images: '/images/logo.png'
+  }
 };
 
-export default ChallengePostPage;
+const page = () => {
+  return <ChallengePostPage />;
+};
+
+export default page;
